refactor: Contest -> Competition 용어 변경
